+++ src/main/java/com/company/project/core/Service.java
@@ -11,7 +11,7 @@ import java.util.List;
 public interface Service<T> {
     void save(T model);//持久化
     void save(List<T> models);//批量持久化
-    void deleteById(Long id);//通过主鍵刪除
+    void deleteById(Integer id);//通过主鍵刪除
     void deleteByIds(String ids);//批量刪除 eg：ids -> “1,2,3,4”
     void update(T model);//更新
     T findById(Long id);//通过ID查找
